project_3_files/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eco-tracker"
version = "0.1.0"
description = "Spec-driven eco tracker social simulator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "typer>=0.12",
  "rich>=13.7"
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
eco = "eco.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
